chore: reduce complexity in controller
