+++ apps/api/src/app/app.controller.ts
@@ -1,9 +1,7 @@
 import { Controller, Get, Param, Post, Body } from '@nestjs/common';
-import { keys } from 'ts-transformer-keys';
 import { Convert2JSON } from '@balance/shared/csv';
 import axios from 'axios';
 import { IPivot, ICreatePivotParams } from '@balance/api-interfaces';
-import { Prisma } from '@prisma/client';
 
 import { PrismaService } from './prisma.service';
 
@@ -18,21 +16,7 @@ export class AppController {
 
   @Get('pivot/:id')
   async getPivotById(@Param('id') id: string): Promise<IPivot> {
-    const IPivotKeys = keys<IPivot>();
-    const IPivotSelectFields: Prisma.PivotSelect = IPivotKeys.reduce(
-      (prev, curr) => {
-        prev[curr] = true;
-        return prev;
-      },
-      {}
-    );
-
-    // eslint-disable-next-line
-    // @ts-ignore
-    return this.dataService.pivot.findUnique({
-      where: { id: Number(id) },
-      select: IPivotSelectFields,
-    });
+    return this.dataService.getPivotById(Number(id));
   }
 
   @Post('pivot')
@@ -40,27 +24,12 @@ export class AppController {
     @Body()
     pivotData: ICreatePivotParams
   ): Promise<number> {
-    // - Fetch URL from `url` var
     const { url, column_name } = pivotData;
     const result = await axios.get(url);
-    const json = Convert2JSON(result.data);
-
-    // grab just Age
-    const pivotArr = json
-      .map((item) => {
-        return item[column_name];
-      })
-      .filter((item) => item)
-      .sort((a, b) => Number(a) - Number(b));
-
-    const formattedAgeArrayString = '[' + pivotArr.join() + ']';
-
-    // INSERT INTO DB
-    const row = await this.dataService.pivot.create({
-      data: {
-        pivot: formattedAgeArrayString,
-      },
-    });
+    const arr = Convert2JSON(result.data, ',', column_name).sort(
+      (a, b) => Number(a) - Number(b)
+    );
+    const row = await this.dataService.createPivotsFromArray(arr);
     return row.id;
   }
 }
