+++ libs/shared/csv/src/lib/shared-csv.ts
@@ -4,7 +4,8 @@ type CSVItem = {
 
 export const Convert2JSON = (
   csv: string,
-  separator = ','
+  separator = ',',
+  column_name = ''
 ): Record<string, string>[] => {
   const csvArr = csv.split('\n');
   const headers: string[] = csvArr[0].split(separator);
@@ -12,10 +13,20 @@ export const Convert2JSON = (
   const rowsWithColumns: any[][] = rows.map((row) => row.split(separator));
 
   const normalize = (str: string) => str.trim().replace(/[\s"]+/g, '');
-  return rowsWithColumns.reduce((jsonArray, row) => {
+  const json = rowsWithColumns.reduce((jsonArray, row) => {
     const item = row.reduce((item: CSVItem[], value: string, index: number) => {
       return { ...item, [normalize(headers[index])]: normalize(value) };
     }, {});
     return jsonArray.concat(item);
   }, []);
+
+  if (column_name) {
+    return json
+      .map((item) => {
+        return item[column_name];
+      })
+      .filter((item) => item);
+  } else {
+    return json;
+  }
 };
